# php-app/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Invoice Management API project",
        "description": "Invoice Management API project. It provides endpoints to manage users, tax profiles, and invoices. The project is structured using the MVC architecture and includes automated tests, Docker support, and API documentation via Postman.",
        "contact": {
            "email": "caiazza.domenico@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/users/{userId}/tax-profiles": {
            "get": {
                "tags": [
                    "TaxProfiles"
                ],
                "summary": "Retrieves user's tax profiles",
                "description": "Returns a paginated collection of tax profiles",
                "operationId": "e6c0d5218d22f58e134dcbd71750942b",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/xApiKeyHeader"
                    },
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "User ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "filter[tax_code]",
                        "in": "query",
                        "description": "Filter for partial search of tax code",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "filter[address]",
                        "in": "query",
                        "description": "Filter for partial address search",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "filter[vat_number]",
                        "in": "query",
                        "description": "Filter for partial VAT number search",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "filter[business_name]",
                        "in": "query",
                        "description": "Filter for partial company name search",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "$ref": "#/components/parameters/per_page"
                    },
                    {
                        "$ref": "#/components/parameters/page"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/TaxProfile"
                                            }
                                        },
                                        "links": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Links"
                                            }
                                        },
                                        "meta": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Meta"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Record not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotFound"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "TaxProfiles"
                ],
                "summary": "Create a new user's tax profile",
                "description": "Creates a new tax profile associated with the specified user.",
                "operationId": "3476ed191bd5ef7a0bb94c4ec5517e61",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/xApiKeyHeader"
                    },
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "User ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Data for creating a TaxProfile",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "tax_code": {
                                        "description": "Tax code (optional)",
                                        "type": "string",
                                        "maxLength": 128
                                    },
                                    "address": {
                                        "description": "Address (optional)",
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "vat_number": {
                                        "description": "VAT number (optional)",
                                        "type": "string",
                                        "maxLength": 128
                                    },
                                    "business_name": {
                                        "description": "Business name (optional)",
                                        "type": "string",
                                        "maxLength": 128
                                    }
                                },
                                "type": "object"
                            },
                            "example": {
                                "tax_code": "RSSMRA80A01F205X",
                                "address": "Via Roma 1, 00100, Milano",
                                "vat_number": "11000500010",
                                "business_name": "Mario Rossi Srl"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "TaxCode created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "TaxProfile created successfully."
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/TaxProfile"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnprocessableEntity"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            }
        },
        "/api/users/{userId}/tax-profiles/{id}": {
            "get": {
                "tags": [
                    "TaxProfiles"
                ],
                "summary": "Get a single TaxProfile",
                "description": "Returns a single TaxProfile by their ID",
                "operationId": "d2051745528b552227e1fed79d7262ba",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/xApiKeyHeader"
                    },
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "ID of user",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the TaxProfile to retrieve",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User getted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TaxProfile"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Record not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotFound"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Links": {
                "properties": {
                    "first": {
                        "type": "string",
                        "example": "[url]?page=1"
                    },
                    "last": {
                        "type": "string",
                        "example": "[url]?page=1"
                    },
                    "prev": {
                        "type": "string",
                        "example": null,
                        "nullable": true
                    },
                    "next": {
                        "type": "string",
                        "example": null,
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "Meta": {
                "properties": {
                    "current_page": {
                        "type": "integer",
                        "example": 1
                    },
                    "from": {
                        "type": "integer",
                        "example": 1
                    },
                    "last_page": {
                        "type": "integer",
                        "example": 1
                    },
                    "links": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "url": {
                                    "type": "string",
                                    "example": null,
                                    "nullable": true
                                },
                                "label": {
                                    "type": "string",
                                    "example": "&laquo; Previous"
                                },
                                "active": {
                                    "type": "boolean",
                                    "example": false
                                }
                            },
                            "type": "object"
                        }
                    },
                    "path": {
                        "type": "string",
                        "example": "http://[path]"
                    },
                    "per_page": {
                        "type": "integer",
                        "example": 10
                    },
                    "to": {
                        "type": "integer",
                        "example": 2
                    },
                    "total": {
                        "type": "integer",
                        "example": 2
                    }
                },
                "type": "object"
            },
            "NotFound": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Record not found"
                    }
                },
                "type": "object"
            },
            "TaxProfile": {
                "required": [
                    "user_id",
                    "tax_code",
                    "address",
                    "vat_number",
                    "business_name"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "tax_code": {
                        "type": "string",
                        "example": "RSSMRA80A01F205X"
                    },
                    "address": {
                        "type": "string",
                        "example": "Via Roma 1, 00100, Milano"
                    },
                    "vat_number": {
                        "type": "string",
                        "example": "11000500010"
                    },
                    "business_name": {
                        "type": "string",
                        "example": "Mario Rossi Srl"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2021-01-01 00:00:00"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2021-01-01 00:00:00"
                    }
                },
                "type": "object"
            },
            "UnprocessableEntity": {
                "properties": {
                    "message": {
                        "properties": {
                            "Field name": {
                                "type": "array",
                                "items": {
                                    "type": "string",
                                    "example": "Field error message"
                                }
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
        },
        "parameters": {
            "per_page": {
                "name": "per_page",
                "in": "query",
                "description": "Results per page",
                "required": false,
                "schema": {
                    "type": "integer",
                    "default": 10
                }
            },
            "page": {
                "name": "page",
                "in": "query",
                "description": "Results page",
                "required": false,
                "schema": {
                    "type": "integer",
                    "default": 1
                }
            },
            "xApiKeyHeader": {
                "name": "x-api-key",
                "in": "header",
                "description": "API key authentication",
                "required": true,
                "schema": {
                    "type": "string",
                    "example": "secret"
                }
            }
        },
        "securitySchemes": {
            "ApiKeyAuth": {
                "type": "apiKey",
                "name": "x-api-key",
                "in": "header"
            }
        }
    },
...
}